move: サイトの根幹となる部分を/coreに移動
