+++ apps/website/build.yaml
@@ -5,3 +5,7 @@ targets:
         options:
           base_locale: ja
           timestamp: false
+          input_directory: lib/core/i18n
+          output_directory: lib/gen/i18n
+          fallback_strategy: base_locale
+          key_case: camel
