+++ apps/website/lib/ui/components/header/site_header.dart
@@ -1,5 +1,5 @@
-import 'package:conference_2024_website/common/util/scroll_to_section.dart';
-import 'package:conference_2024_website/i18n/strings.g.dart';
+import 'package:conference_2024_website/core/util/scroll_to_section.dart';
+import 'package:conference_2024_website/core/i18n/strings.g.dart';
 import 'package:conference_2024_website/ui/theme/extension/theme_extension.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
