+++ apps/website/lib/ui/home/components/news_component.dart
@@ -2,7 +2,7 @@ import 'dart:async';
 
 import 'package:collection/collection.dart';
 import 'package:common_data/news.dart';
-import 'package:conference_2024_website/i18n/strings.g.dart';
+import 'package:conference_2024_website/core/i18n/strings.g.dart';
 import 'package:conference_2024_website/ui/home/notifier/news_data_notifier.dart';
 import 'package:conference_2024_website/ui/home/notifier/news_data_state.dart';
 import 'package:conference_2024_website/ui/theme/extension/theme_extension.dart';
