+++ apps/website/lib/ui/home/components/speaker_wanted.dart
@@ -1,4 +1,4 @@
-import 'package:conference_2024_website/i18n/strings.g.dart';
+import 'package:conference_2024_website/core/i18n/strings.g.dart';
 import 'package:conference_2024_website/ui/components/button/app_button.dart';
 import 'package:conference_2024_website/ui/theme/extension/theme_extension.dart';
 import 'package:flutter/material.dart';
