+++ apps/website/lib/ui/home/components/title_and_logo.dart
@@ -1,4 +1,4 @@
-import 'package:conference_2024_website/i18n/strings.g.dart';
+import 'package:conference_2024_website/core/i18n/strings.g.dart';
 import 'package:conference_2024_website/ui/theme/extension/theme_extension.dart';
 import 'package:flutter/material.dart';
